requirements.txt
Django~=5.1.3
djangorestframework-simplejwt~=5.3.1
djangorestframework~=3.15.2
future~=1.0.0
requests~=2.25.1
beautifulsoup4~=4.10.0